+++ ouds_theme_contract/lib/ouds_tokens_provider.dart
@@ -14,6 +14,7 @@ import 'package:ouds_theme_contract/ouds_color_scheme.dart';
 import 'package:ouds_theme_contract/theme/tokens/semantic/ouds_border_semantic_tokens.dart';
 import 'package:ouds_theme_contract/theme/tokens/semantic/ouds_elevation_semantic_tokens.dart';
 import 'package:ouds_theme_contract/theme/tokens/semantic/ouds_font_semantic_tokens.dart';
+import 'package:ouds_theme_contract/theme/tokens/semantic/ouds_grid_semantic_tokens.dart';
 import 'package:ouds_theme_contract/theme/tokens/semantic/ouds_opacity_semantic_tokens.dart';
 import 'package:ouds_theme_contract/theme/tokens/semantic/ouds_size_semantic_tokens.dart';
 import 'package:ouds_theme_contract/theme/tokens/semantic/ouds_space_semantic_tokens.dart';
@@ -30,6 +31,7 @@ class OudsProvidersTokens {
   final OudsElevationSemanticTokens elevationTokens;
   final OudsSpaceSemanticTokens spaceTokens;
   final OudsSizeSemanticTokens sizeTokens;
+  final OudsGridSemanticTokens gridTokens;
   final OudsFontSemanticTokens fontTokens;
 
   OudsProvidersTokens({
@@ -39,6 +41,7 @@ class OudsProvidersTokens {
     required this.elevationTokens,
     required this.spaceTokens,
     required this.sizeTokens,
+    required this.gridTokens,
     required this.fontTokens,
   });
 }